x86: normalize types used with _mm{,256}_set?_epi16
w/clang -fsanitize=integer fixes warnings of the form:
implicit conversion from type 'uint16_t' (aka 'unsigned short') of value
49151 (16-bit, unsigned) to type 'short' changed the value to -16385
(16-bit, signed)

Bug: aomedia:3136
Bug: b/229626362

+++ aom_dsp/simd/v128_intrinsics_x86.h
@@ -83,7 +83,7 @@ SIMD_INLINE v128 v128_zero() { return _mm_setzero_si128(); }
 
 SIMD_INLINE v128 v128_dup_8(uint8_t x) { return _mm_set1_epi8((char)x); }
 
-SIMD_INLINE v128 v128_dup_16(uint16_t x) { return _mm_set1_epi16(x); }
+SIMD_INLINE v128 v128_dup_16(uint16_t x) { return _mm_set1_epi16((short)x); }
 
 SIMD_INLINE v128 v128_dup_32(uint32_t x) { return _mm_set1_epi32(x); }
